Original File Upload
As I continue to do an overhaul of my mIRC scripts, I wrote this simple alias script to mess with certain Eggdrop bots. The original script I wrote years ago only had about 6 items to get an Eggdrop bot high. I cannot remember the name of the dependency needed on the TCL side of things, but there is one needed in order for this script to work properly. Anyone can change this one to how they like.